// src/lib/api-client/models/api-create-keyword-set-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Domain Flow API
 * A comprehensive API for domain generation, validation, and management operations.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ApiKeywordRuleRequest } from './api-keyword-rule-request';

/**
 * 
 * @export
 * @interface ApiCreateKeywordSetRequest
 */
export interface ApiCreateKeywordSetRequest {
    /**
     * 
     * @type {string}
     * @memberof ApiCreateKeywordSetRequest
     */
    'description'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ApiCreateKeywordSetRequest
     */
    'isEnabled'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ApiCreateKeywordSetRequest
     */
    'name': string;
    /**
     * 
     * @type {Array<ApiKeywordRuleRequest>}
     * @memberof ApiCreateKeywordSetRequest
     */
    'rules'?: Array<ApiKeywordRuleRequest>;
}