// src/lib/api-client/models/api-update-keyword-set-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Domain Flow API
 * A comprehensive API for domain generation, validation, and management operations.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ApiKeywordRuleRequest } from './api-keyword-rule-request';

/**
 * 
 * @export
 * @interface ApiUpdateKeywordSetRequest
 */
export interface ApiUpdateKeywordSetRequest {
    /**
     * 
     * @type {string}
     * @memberof ApiUpdateKeywordSetRequest
     */
    'description'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ApiUpdateKeywordSetRequest
     */
    'isEnabled'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ApiUpdateKeywordSetRequest
     */
    'name'?: string;
    /**
     * 
     * @type {Array<ApiKeywordRuleRequest>}
     * @memberof ApiUpdateKeywordSetRequest
     */
    'rules'?: Array<ApiKeywordRuleRequest>;
}